+++ docs/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 * Notification for stop, to provide feedback
 
+### Changed
+* Spinner color to orange
+
 ### Fixed
 * Improve some slow views
   * Subjects lists